+++ src/handlers/validator.rs
@@ -21,11 +21,11 @@
 
 use crate::cache::{get_conn, RedisPool};
 use crate::errors::{ApiError, CacheError};
-use crate::helpers::{respond_json};
+use crate::helpers::respond_json;
 use crate::sync::sync::Key;
 use actix_web::web::{Data, Json, Path, Query};
 use redis::aio::Connection;
-use serde::{Deserialize, Serialize};
+use serde::{de::Deserializer, Deserialize, Serialize};
 use std::{collections::BTreeMap, str::FromStr};
 use substrate_subxt::{sp_runtime::AccountId32, staking::EraIndex};
 
@@ -33,10 +33,16 @@ type ValidatorCache = BTreeMap<String, String>;
 type ValidatorEraCache = BTreeMap<String, String>;
 
 #[derive(Debug, Serialize, PartialEq)]
-pub struct ValidatorResponse {
+pub struct Validator {
     pub stash: String,
     pub controller: String,
     pub name: String,
+    pub own_stake: u128,
+    pub total_stake: u128,
+    pub others_stake: u128,
+    pub stakers: u32,
+    pub others_stake_clipped: u128,
+    pub stakers_clipped: u32,
     pub inclusion_rate: f32,
     pub mean_reward_points: f64,
     pub commission: u32,
@@ -45,16 +51,46 @@ pub struct ValidatorResponse {
     pub reward_staked: bool,
 }
 
-impl From<ValidatorCache> for ValidatorResponse {
+impl From<ValidatorCache> for Validator {
     fn from(data: ValidatorCache) -> Self {
         let zero = "0".to_string();
-        ValidatorResponse {
+        Validator {
             stash: data.get("stash").unwrap_or(&"".to_string()).to_string(),
             controller: data
                 .get("controller")
                 .unwrap_or(&"".to_string())
                 .to_string(),
             name: data.get("name").unwrap_or(&"".to_string()).to_string(),
+            own_stake: data
+                .get("own_stake")
+                .unwrap_or(&zero)
+                .parse::<u128>()
+                .unwrap_or_default(),
+            total_stake: data
+                .get("total_stake")
+                .unwrap_or(&zero)
+                .parse::<u128>()
+                .unwrap_or_default(),
+            others_stake: data
+                .get("others_stake")
+                .unwrap_or(&zero)
+                .parse::<u128>()
+                .unwrap_or_default(),
+            stakers: data
+                .get("stakers")
+                .unwrap_or(&zero)
+                .parse::<u32>()
+                .unwrap_or_default(),
+            others_stake_clipped: data
+                .get("others_stake_clipped")
+                .unwrap_or(&zero)
+                .parse::<u128>()
+                .unwrap_or_default(),
+            stakers_clipped: data
+                .get("stakers_clipped")
+                .unwrap_or(&zero)
+                .parse::<u32>()
+                .unwrap_or_default(),
             inclusion_rate: data
                 .get("inclusion_rate")
                 .unwrap_or(&zero)
@@ -89,6 +125,8 @@ impl From<ValidatorCache> for ValidatorResponse {
     }
 }
 
+type ValidatorResponse = Validator;
+
 /// Get a validator
 pub async fn get_validator(
     stash: Path<String>,
@@ -116,6 +154,9 @@ pub struct ValidatorEra {
     pub own_stake: u128,
     pub total_stake: u128,
     pub others_stake: u128,
+    pub stakers: u32,
+    pub others_stake_clipped: u128,
+    pub stakers_clipped: u32,
     pub reward_points: u32,
     pub commission: u32,
     pub blocked: bool,
@@ -151,6 +192,21 @@ impl From<ValidatorEraCache> for ValidatorEra {
                 .get("others_stake")
                 .unwrap_or(&zero)
                 .parse::<u128>()
+                .unwrap_or_default(),
+            stakers: data
+                .get("stakers")
+                .unwrap_or(&zero)
+                .parse::<u32>()
+                .unwrap_or_default(),
+            others_stake_clipped: data
+                .get("others_stake_clipped")
+                .unwrap_or(&zero)
+                .parse::<u128>()
+                .unwrap_or_default(),
+            stakers_clipped: data
+                .get("stakers_clipped")
+                .unwrap_or(&zero)
+                .parse::<u32>()
                 .unwrap_or_default(),
             reward_points: data
                 .get("reward_points")
@@ -233,20 +289,153 @@ pub async fn get_validator_eras(
     })
 }
 
-#[derive(Debug, Deserialize, PartialEq)]
+#[derive(Debug, Deserialize, Clone, PartialEq)]
 enum Queries {
     All = 1,
     Active = 2,
+    Board = 3,
 }
 
-#[derive(Debug, Deserialize)]
+impl std::fmt::Display for Queries {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            Self::All => write!(f, "all"),
+            Self::Active => write!(f, "active"),
+            Self::Board => write!(f, "board"),
+        }
+    }
+}
+
+/// Weight can be any value in a 10-point scale. Higher the weight more important
+/// is the criteria to the user
+type Weight = u32;
+
+/// Weights represent an array of points, where the points in each position represents
+/// the weight for the respective criteria
+/// Position 0 - Higher Inclusion rate is preferrable
+/// Position 1 - Lower Commission is preferrable
+/// Position 2 - if reward is staked is preferrable
+/// Position 3 - if in active set is preferrable
+type Weights = Vec<Weight>;
+
+// Number of elements to return
+type Quantity = u32;
+
+#[derive(Debug, Deserialize, Clone, PartialEq)]
 pub struct Params {
     q: Queries,
+    #[serde(default)]
+    #[serde(deserialize_with = "parse_weights")]
+    w: Weights,
+    n: Quantity,
+}
+
+fn parse_weights<'de, D>(d: D) -> Result<Weights, D::Error>
+where
+    D: Deserializer<'de>,
+{
+    Deserialize::deserialize(d).map(|x: Option<_>| {
+        let weights_str = x.unwrap_or("".to_string());
+
+        let weights_vs: Vec<&str> = weights_str.split(",").collect();
+
+        weights_vs
+            .into_iter()
+            .map(|y| {
+                let weight: u32 = String::from(y).parse().unwrap_or(5);
+                if weight > 10 {
+                    return 10;
+                }
+                weight
+            })
+            .collect()
+    })
 }
 
 #[derive(Debug, Serialize, PartialEq)]
 pub struct ValidatorsResponse {
     pub data: Vec<String>,
+}
+
+fn get_board_name(weights: &Weights) -> String {
+    weights
+        .iter()
+        .enumerate()
+        .map(|(i, x)| {
+            if i == 0 {
+                return x.to_string();
+            }
+            format!(",{}", x)
+        })
+        .collect()
+}
+
+fn normaliza_inclusion(inclusion_rate: f32) -> u32 {
+    let r = inclusion_rate * 100.0;
+    r.round() as u32
+}
+
+fn normaliza_commission(commission: u32) -> u32 {
+    100 - (commission / 10000000)
+}
+
+async fn generate_board(
+    era_index: EraIndex,
+    weights: &Weights,
+    cache: Data<RedisPool>,
+) -> Result<(), ApiError> {
+    let mut conn = get_conn(&cache).await?;
+
+    let stashes: Vec<String> = redis::cmd("ZRANGE")
+        .arg(Key::BoardAtEra(era_index, Queries::All.to_string()))
+        .arg("-inf")
+        .arg("+inf")
+        .arg("BYSCORE")
+        .query_async(&mut conn as &mut Connection)
+        .await
+        .map_err(CacheError::RedisCMDError)?;
+
+    let key = Key::BoardAtEra(era_index, get_board_name(weights));
+    for stash in stashes {
+        let stash = AccountId32::from_str(&*stash.to_string())?;
+        let data: ValidatorCache = redis::cmd("HGETALL")
+            .arg(Key::Validator(stash.clone()))
+            .query_async(&mut conn as &mut Connection)
+            .await
+            .map_err(CacheError::RedisCMDError)?;
+
+        let validator: Validator = data.into();
+        // If the validator does not accept nominations
+        // score is not given
+        if validator.blocked {
+            continue;
+        }
+
+        let score = normaliza_inclusion(validator.inclusion_rate) * weights[0]
+            + normaliza_commission(validator.commission) * weights[1]
+            + validator.reward_staked as u32 * weights[2] * 100
+            + validator.active as u32 * weights[3] * 100;
+
+        println!(
+            "{} * {} + {} * {} = {} -> {}",
+            normaliza_inclusion(validator.inclusion_rate),
+            weights[0],
+            normaliza_commission(validator.commission),
+            weights[1],
+            score,
+            stash.to_string()
+        );
+
+        let _: () = redis::cmd("ZADD")
+            .arg(key.to_string())
+            .arg(score) // score
+            .arg(stash.to_string()) // member
+            .query_async(&mut conn as &mut Connection)
+            .await
+            .map_err(CacheError::RedisCMDError)?;
+    }
+
+    Ok(())
 }
 
 /// Get a validators
@@ -262,16 +451,34 @@ pub async fn get_validators(
         .map_err(CacheError::RedisCMDError)?;
 
     let key = match params.q {
-        Queries::Active => Key::AllValidatorsByEra(era_index),
-        Queries::All => Key::ActiveValidatorsByEra(era_index),
+        Queries::Active => Key::BoardAtEra(era_index, Queries::Active.to_string()),
+        Queries::All => Key::BoardAtEra(era_index, Queries::All.to_string()),
+        Queries::Board => Key::BoardAtEra(era_index, get_board_name(&params.w)),
     };
-    let stashes: Vec<String> = redis::cmd("SMEMBERS")
-        .arg(key)
+
+    let exists: bool = redis::cmd("EXISTS")
+        .arg(key.clone())
         .query_async(&mut conn as &mut Connection)
         .await
         .map_err(CacheError::RedisCMDError)?;
 
-    respond_json(ValidatorsResponse {
-        data: stashes,
-    })
+    if !exists {
+        // Generate and cache leaderboard
+        generate_board(era_index, &params.w, cache).await?;
+    }
+
+    let stashes: Vec<String> = redis::cmd("ZRANGE")
+        .arg(key.clone())
+        .arg("+inf")
+        .arg("0")
+        .arg("BYSCORE")
+        .arg("REV")
+        .arg("LIMIT")
+        .arg("0")
+        .arg(params.n)
+        .query_async(&mut conn as &mut Connection)
+        .await
+        .map_err(CacheError::RedisCMDError)?;
+
+    respond_json(ValidatorsResponse { data: stashes })
 }