+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Generate boards from user defined weights
+- Change active and all validators set to sorted set
+- Add stakers clipped
+- Add number of stakers and clipped stakers
+
 ## [0.1.0] - 2021-04-19
 
 - First release